api_entite_micro_service/requirements.txt
flask
flask_restx
kafka-python